requirements.txt
django
django-environ
djangorestframework
requests
django-cors-headers
drf-yasg
Pillow
